requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
firebase-admin
altair